Refactor application_handler into services layer with db_gateways file.
